+++ vllm/entrypoints/generate/generative_scoring/api_router.py
@@ -1,8 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-from http import HTTPStatus
 
-from fastapi import APIRouter, Depends, FastAPI, Request
+from fastapi import Request
 from fastapi.responses import JSONResponse
 
 from vllm.entrypoints.generate.generative_scoring.serving import (
@@ -12,12 +11,9 @@ from vllm.entrypoints.generate.generative_scoring.serving import (
 from vllm.entrypoints.openai.engine.protocol import ErrorResponse
 from vllm.entrypoints.serve.utils.api_utils import (
     load_aware_call,
-    validate_json_request,
     with_cancellation,
 )
 from vllm.logger import init_logger
-
-router = APIRouter()
 
 logger = init_logger(__name__)
 
@@ -26,14 +22,6 @@ def generative_scoring(request: Request) -> ServingGenerativeScoring | None:
     return request.app.state.serving_generative_scoring
 
 
-@router.post(
-    "/generative_scoring",
-    dependencies=[Depends(validate_json_request)],
-    responses={
-        HTTPStatus.BAD_REQUEST.value: {"model": ErrorResponse},
-        HTTPStatus.INTERNAL_SERVER_ERROR.value: {"model": ErrorResponse},
-    },
-)
 @with_cancellation
 @load_aware_call
 async def create_generative_scoring(raw_request: Request):
@@ -58,7 +46,3 @@ async def create_generative_scoring(raw_request: Request):
         return JSONResponse(content=result.model_dump())
 
     raise ValueError(f"Unexpected response type: {type(result)}")
-
-
-def register_generative_scoring_api_router(app: FastAPI):
-    app.include_router(router)
