+++ vllm/entrypoints/pooling/classify/api_router.py
@@ -1,26 +1,22 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
-from fastapi import APIRouter, Depends, Request
+from fastapi import Request
 from fastapi.responses import Response
 
 from vllm.entrypoints.serve.utils.api_utils import (
     load_aware_call,
-    validate_json_request,
     with_cancellation,
 )
 
 from .protocol import ClassificationRequest
 from .serving import ServingClassification
 
-router = APIRouter()
-
 
 def classify(request: Request) -> ServingClassification | None:
     return request.app.state.serving_classification
 
 
-@router.post("/classify", dependencies=[Depends(validate_json_request)])
 @with_cancellation
 @load_aware_call
 async def create_classify(
