+++ vllm/entrypoints/serve/dev/server_info/api_router.py
@@ -7,7 +7,7 @@ import functools
 from typing import Annotated, Literal
 
 import pydantic
-from fastapi import APIRouter, FastAPI, Query, Request
+from fastapi import Query, Request
 from fastapi.responses import JSONResponse
 
 import vllm.envs as envs
@@ -17,8 +17,6 @@ from vllm.logger import init_logger
 
 logger = init_logger(__name__)
 
-
-router = APIRouter()
 PydanticVllmConfig = pydantic.TypeAdapter(VllmConfig)
 
 
@@ -40,7 +38,6 @@ def _get_system_env_info_cached():
     return get_env_info()._asdict()
 
 
-@router.get("/server_info")
 async def show_server_info(
     raw_request: Request,
     config_format: Annotated[Literal["text", "json"], Query()] = "text",
@@ -57,7 +54,3 @@ async def show_server_info(
         "system_env": await asyncio.to_thread(_get_system_env_info_cached),
     }
     return JSONResponse(content=server_info)
-
-
-def attach_router(app: FastAPI):
-    app.include_router(router)
