+++ vllm/entrypoints/serve/instrumentator/basic.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
-from fastapi import APIRouter, Request
+from fastapi import Request
 from fastapi.responses import JSONResponse
 
 from vllm.engine.protocol import EngineClient
@@ -9,8 +9,6 @@ from vllm.entrypoints.openai.engine.serving import OpenAIServing
 from vllm.entrypoints.serve.tokenize.serving import OpenAIServingTokenization
 from vllm.logger import init_logger
 from vllm.version import __version__ as VLLM_VERSION
-
-router = APIRouter()
 
 logger = init_logger(__name__)
 
@@ -28,7 +26,6 @@ def engine_client(request: Request) -> EngineClient:
     return request.app.state.engine_client
 
 
-@router.get("/load")
 async def get_server_load_metrics(request: Request):
     # This endpoint returns the current server load metrics.
     # It tracks requests utilizing the GPU from the following routes:
@@ -51,7 +48,6 @@ async def get_server_load_metrics(request: Request):
     return JSONResponse(content={"server_load": request.app.state.server_load_metrics})
 
 
-@router.get("/version")
 async def show_version():
     ver = {"version": VLLM_VERSION}
     return JSONResponse(content=ver)
