+++ vllm/entrypoints/serve/instrumentator/health.py
@@ -2,7 +2,7 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
 
-from fastapi import APIRouter, Request
+from fastapi import Request
 from fastapi.responses import Response
 
 from vllm.engine.protocol import EngineClient
@@ -12,14 +12,10 @@ from vllm.v1.engine.exceptions import EngineDeadError
 logger = init_logger(__name__)
 
 
-router = APIRouter()
-
-
 def engine_client(request: Request) -> EngineClient:
     return request.app.state.engine_client
 
 
-@router.get("/health", response_class=Response)
 async def health(raw_request: Request) -> Response:
     """Health check."""
     client = engine_client(raw_request)
