+++ vllm/entrypoints/serve/render/api_router.py
@@ -1,48 +1,26 @@
 # SPDX-License-Identifier: Apache-2.0
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
-from http import HTTPStatus
 
-from fastapi import APIRouter, Depends, FastAPI, Request
+from fastapi import Request
 from fastapi.responses import JSONResponse
 
-from vllm.entrypoints.openai.chat_completion.protocol import (
-    ChatCompletionRequest,
-    ChatCompletionResponse,
-)
-from vllm.entrypoints.openai.completion.protocol import (
-    CompletionRequest,
-    CompletionResponse,
-)
+from vllm.entrypoints.openai.chat_completion.protocol import ChatCompletionRequest
+from vllm.entrypoints.openai.completion.protocol import CompletionRequest
 from vllm.entrypoints.openai.engine.protocol import ErrorResponse
 from vllm.entrypoints.serve.disagg.protocol import (
     DerenderChatRequest,
     DerenderCompletionRequest,
-    GenerateRequest,
 )
 from vllm.entrypoints.serve.render.serving import OpenAIServingRender
-from vllm.entrypoints.serve.utils.api_utils import validate_json_request
 from vllm.logger import init_logger
 
 logger = init_logger(__name__)
-
-router = APIRouter()
 
 
 def render(request: Request) -> OpenAIServingRender | None:
     return getattr(request.app.state, "openai_serving_render", None)
 
 
-@router.post(
-    "/v1/chat/completions/render",
-    dependencies=[Depends(validate_json_request)],
-    response_model=GenerateRequest,
-    responses={
-        HTTPStatus.BAD_REQUEST.value: {"model": ErrorResponse},
-        HTTPStatus.NOT_FOUND.value: {"model": ErrorResponse},
-        HTTPStatus.NOT_IMPLEMENTED.value: {"model": ErrorResponse},
-        HTTPStatus.INTERNAL_SERVER_ERROR.value: {"model": ErrorResponse},
-    },
-)
 async def render_chat_completion(request: ChatCompletionRequest, raw_request: Request):
     handler = render(raw_request)
     if handler is None:
@@ -58,16 +36,6 @@ async def render_chat_completion(request: ChatCompletionRequest, raw_request: Re
     return JSONResponse(content=result.model_dump())
 
 
-@router.post(
-    "/v1/completions/render",
-    dependencies=[Depends(validate_json_request)],
-    response_model=list[GenerateRequest],
-    responses={
-        HTTPStatus.BAD_REQUEST.value: {"model": ErrorResponse},
-        HTTPStatus.NOT_FOUND.value: {"model": ErrorResponse},
-        HTTPStatus.INTERNAL_SERVER_ERROR.value: {"model": ErrorResponse},
-    },
-)
 async def render_completion(request: CompletionRequest, raw_request: Request):
     handler = render(raw_request)
     if handler is None:
@@ -81,16 +49,6 @@ async def render_completion(request: CompletionRequest, raw_request: Request):
     return JSONResponse(content=[item.model_dump() for item in result])
 
 
-@router.post(
-    "/v1/chat/completions/derender",
-    dependencies=[Depends(validate_json_request)],
-    response_model=ChatCompletionResponse,
-    responses={
-        HTTPStatus.BAD_REQUEST.value: {"model": ErrorResponse},
-        HTTPStatus.NOT_FOUND.value: {"model": ErrorResponse},
-        HTTPStatus.INTERNAL_SERVER_ERROR.value: {"model": ErrorResponse},
-    },
-)
 async def derender_chat_completion(request: DerenderChatRequest, raw_request: Request):
     handler = render(raw_request)
     if handler is None:
@@ -106,16 +64,6 @@ async def derender_chat_completion(request: DerenderChatRequest, raw_request: Re
     return JSONResponse(content=result.model_dump())
 
 
-@router.post(
-    "/v1/completions/derender",
-    dependencies=[Depends(validate_json_request)],
-    response_model=CompletionResponse,
-    responses={
-        HTTPStatus.BAD_REQUEST.value: {"model": ErrorResponse},
-        HTTPStatus.NOT_FOUND.value: {"model": ErrorResponse},
-        HTTPStatus.INTERNAL_SERVER_ERROR.value: {"model": ErrorResponse},
-    },
-)
 async def derender_completion(request: DerenderCompletionRequest, raw_request: Request):
     handler = render(raw_request)
     if handler is None:
@@ -127,7 +75,3 @@ async def derender_completion(request: DerenderCompletionRequest, raw_request: R
         return JSONResponse(content=result.model_dump(), status_code=result.error.code)
 
     return JSONResponse(content=result.model_dump())
-
-
-def attach_router(app: FastAPI) -> None:
-    app.include_router(router)
