+++ vllm/entrypoints/serve/sagemaker/api_router.py
@@ -7,18 +7,16 @@ from typing import Any
 
 import model_hosting_container_standards.sagemaker as sagemaker_standards
 import pydantic
-from fastapi import APIRouter, Depends, FastAPI, HTTPException, Request
+from fastapi import FastAPI, HTTPException, Request
 from fastapi.responses import JSONResponse, Response
 
 from vllm.config import ModelConfig
 from vllm.entrypoints.generate.factories import get_generate_invocation_types
-from vllm.entrypoints.openai.engine.protocol import ErrorResponse
 from vllm.entrypoints.openai.engine.serving import OpenAIServing
 from vllm.entrypoints.pooling.base.serving import PoolingServingBase
 from vllm.entrypoints.pooling.factories import get_pooling_invocation_types
 from vllm.entrypoints.serve.instrumentator.basic import base
 from vllm.entrypoints.serve.instrumentator.health import health
-from vllm.entrypoints.serve.utils.api_utils import validate_json_request
 from vllm.tasks import SupportedTask
 
 # TODO: RequestType = TypeForm[BaseModel] when recognized by type checkers
@@ -28,13 +26,16 @@ GetHandlerFn = Callable[[Request], OpenAIServing | PoolingServingBase | None]
 EndpointFn = Callable[[RequestType, Request], Awaitable[Any]]
 
 
-def attach_router(
-    app: FastAPI,
+@sagemaker_standards.register_ping_handler
+async def ping(raw_request: Request) -> Response:
+    """Ping check. Endpoint required for SageMaker"""
+    return await health(raw_request)
+
+
+def invocations(
     supported_tasks: tuple["SupportedTask", ...],
     model_config: ModelConfig | None = None,
 ):
-    router = APIRouter()
-
     # NOTE: Construct the TypeAdapters only once
     INVOCATION_TYPES = get_generate_invocation_types(
         supported_tasks, model_config
@@ -45,26 +46,10 @@ def attach_router(
         for request_type, (get_handler, endpoint) in INVOCATION_TYPES
     ]
 
-    @router.post("/ping", response_class=Response)
-    @router.get("/ping", response_class=Response)
-    @sagemaker_standards.register_ping_handler
-    async def ping(raw_request: Request) -> Response:
-        """Ping check. Endpoint required for SageMaker"""
-        return await health(raw_request)
-
-    @router.post(
-        "/invocations",
-        dependencies=[Depends(validate_json_request)],
-        responses={
-            HTTPStatus.BAD_REQUEST.value: {"model": ErrorResponse},
-            HTTPStatus.UNSUPPORTED_MEDIA_TYPE.value: {"model": ErrorResponse},
-            HTTPStatus.INTERNAL_SERVER_ERROR.value: {"model": ErrorResponse},
-        },
-    )
     @sagemaker_standards.register_invocation_handler
     @sagemaker_standards.stateful_session_manager()
     @sagemaker_standards.inject_adapter_id(adapter_path="model")
-    async def invocations(raw_request: Request):
+    async def _invocations(raw_request: Request):
         """For SageMaker, routes requests based on the request type."""
         try:
             body = await raw_request.json()
@@ -96,7 +81,7 @@ def attach_router(
         res = base(raw_request).create_error_response(message=msg)
         return JSONResponse(content=res.model_dump(), status_code=res.error.code)
 
-    app.include_router(router)
+    return _invocations
 
 
 def sagemaker_standards_bootstrap(app: FastAPI) -> FastAPI:
